+++ opal/mca/threads/base/Makefile.am
@@ -22,5 +22,6 @@ headers += \
         base/base.h
 
 libmca_threads_la_SOURCES += \
+        base/create_join.c \
         base/threads_base.c \
         base/tsd.c
