+++ opal/mca/threads/base/create_join.c
@@ -24,6 +24,7 @@
  */
 
 #include <unistd.h>
+#include <pthread.h>
 
 #include "opal/constants.h"
 #include "opal/mca/threads/threads.h"
@@ -32,9 +33,51 @@
 #include "opal/util/output.h"
 #include "opal/util/sys_limits.h"
 
-int opal_tsd_key_create(opal_tsd_key_t *key, opal_tsd_destructor_t destructor)
+/*
+ * Constructor
+ */
+static void opal_thread_construct(opal_thread_t *t)
+{
+    t->t_run = 0;
+    t->t_handle = (pthread_t) -1;
+}
+
+OBJ_CLASS_INSTANCE(opal_thread_t, opal_object_t, opal_thread_construct, NULL);
+
+int opal_thread_start(opal_thread_t *t)
 {
     int rc;
-    rc = pthread_key_create(key, destructor);
+
+    if (OPAL_ENABLE_DEBUG) {
+        if (NULL == t->t_run || (pthread_t) -1 != t->t_handle) {
+            return OPAL_ERR_BAD_PARAM;
+        }
+    }
+
+    rc = pthread_create(&t->t_handle, NULL, (void *(*) (void *) ) t->t_run, t);
+
     return 0 == rc ? OPAL_SUCCESS : OPAL_ERR_IN_ERRNO;
+}
+
+int opal_thread_join(opal_thread_t *t, void **thr_return)
+{
+    int rc = pthread_join(t->t_handle, thr_return);
+    t->t_handle = (pthread_t) -1;
+    return 0 == rc ? OPAL_SUCCESS : OPAL_ERR_IN_ERRNO;
+}
+
+bool opal_thread_self_compare(opal_thread_t *t)
+{
+    return pthread_self() == t->t_handle;
+}
+
+opal_thread_t *opal_thread_get_self(void)
+{
+    opal_thread_t *t = OBJ_NEW(opal_thread_t);
+    t->t_handle = pthread_self();
+    return t;
+}
+
+void opal_thread_set_main(void)
+{
 }
