+++ opal/mca/threads/qthreads/threads_qthreads_module.c
@@ -3,7 +3,7 @@
  * Copyright (c) 2004-2005 The Trustees of Indiana University and Indiana
  *                         University Research and Technology
  *                         Corporation.  All rights reserved.
- * Copyright (c) 2004-2005 The University of Tennessee and The University
+ * Copyright (c) 2004-2021 The University of Tennessee and The University
  *                         of Tennessee Research Foundation.  All rights
  *                         reserved.
  * Copyright (c) 2004-2005 High Performance Computing Center Stuttgart,
@@ -50,74 +50,6 @@ static inline void self_key_ensure_init(void)
     }
     /* opal_thread_self_key has been already initialized. */
 }
-
-/*
- * Constructor
- */
-static void opal_thread_construct(opal_thread_t *t)
-{
-    t->t_run = 0;
-    t->t_thread_ret = 0;
-}
-
-OBJ_CLASS_INSTANCE(opal_thread_t, opal_object_t, opal_thread_construct, NULL);
-
-static inline aligned_t *opal_thread_get_qthreads_self(void)
-{
-    self_key_ensure_init();
-    void *ptr = qthread_getspecific(opal_thread_self_key);
-    return (aligned_t *) ptr;
-}
-
-static aligned_t opal_thread_qthreads_wrapper(void *arg)
-{
-    opal_thread_t *t = (opal_thread_t *) arg;
-
-    /* Register itself. */
-    self_key_ensure_init();
-    qthread_setspecific(opal_thread_self_key, t->t_thread_ret_ptr);
-
-    t->t_ret = ((void *(*) (void *) ) t->t_run)(t);
-    return 0;
-}
-
-opal_thread_t *opal_thread_get_self(void)
-{
-    opal_threads_ensure_init_qthreads();
-    opal_thread_t *t = OBJ_NEW(opal_thread_t);
-    t->t_thread_ret_ptr = opal_thread_get_qthreads_self();
-    return t;
-}
-
-bool opal_thread_self_compare(opal_thread_t *t)
-{
-    opal_threads_ensure_init_qthreads();
-    return opal_thread_get_qthreads_self() == &t->t_thread_ret;
-}
-
-int opal_thread_join(opal_thread_t *t, void **thr_return)
-{
-    qthread_readFF(NULL, t->t_thread_ret_ptr);
-    if (thr_return) {
-        *thr_return = t->t_ret;
-    }
-    t->t_thread_ret = 0;
-    return OPAL_SUCCESS;
-}
-
-void opal_thread_set_main(void)
-{
-}
-
-int opal_thread_start(opal_thread_t *t)
-{
-    opal_threads_ensure_init_qthreads();
-    t->t_thread_ret_ptr = &t->t_thread_ret;
-    qthread_fork(opal_thread_qthreads_wrapper, t, &t->t_thread_ret);
-    return OPAL_SUCCESS;
-}
-
-OBJ_CLASS_DECLARATION(opal_thread_t);
 
 int opal_tsd_key_create(opal_tsd_key_t *key, opal_tsd_destructor_t destructor)
 {
